refactor: config in seperate file/object
